HDFS-12698. Ozone: Use time units in the Ozone configuration values. Contributed by Elek, Marton.

+++ hadoop-hdfs-project/hadoop-hdfs/src/main/java/org/apache/hadoop/ozone/ksm/KeyDeletingService.java
@@ -57,7 +57,7 @@ public class KeyDeletingService extends BackgroundService {
   private final int keyLimitPerTask;
 
   public KeyDeletingService(ScmBlockLocationProtocol scmClient,
-      KeyManager manager, int serviceInterval,
+      KeyManager manager, long serviceInterval,
       long serviceTimeout, Configuration conf) {
     super("KeyDeletingService", serviceInterval, TimeUnit.MILLISECONDS,
         KEY_DELETING_CORE_POOL_SIZE, serviceTimeout);
